Fixes long-standing movement-in-collider bug.
If the viewer moved into a collider and then immediately reversed
direction, they could get the cards into an off-by-one situation
where no card is visible. This adds almost the same behavior to
OnTriggerExit as OnTriggerEnter. Now card is chosen on both entry
and exit of the card's collider to make sure it's always right.